07/28/2022 StringBuilder example code.  Also updated the code for gathering text one char at a time.
Note: Using StringBuilder in the code to read a character at a time produced all caps, an underscore for a period, and other strange output.  I believe it has to do with how StringBuilder and Console.ReadKey() work together.  Or maybe it is the Convert.ToChar(key.KeyChar) method that does not work with StringBuilders.