08/12/2022 Functions.
Found out that you can indeed pass by reference with C#.  Keyword ref.  Must be explicit in function definition and function call.

Some other simple function examples.

Also, saw that using System; only needs to be in one project in the whole solution.  That might be a bit confusing for students in the beginning.  But it is there right from the start of the first project when the solution is made so I can point it out if they are doing as I suggest and make projects inside their one solution.